requirements.txt
pandas>=2.0.0
python-dateutil>=2.8.0
pytest>=7.0.0

